Load from and store people in `people.json` file
