+++ store/books.json
@@ -1,1 +1,1 @@
-[]
+[{"title":"fdsfs","author":"fdsff"},{"title":"fdsfs","author":"fdsfd"}]
