+++ store/people.json
@@ -1,1 +1,1 @@
-[]
+[{"class":"Student","name":"Unknown","age":"chris","id":96}]
